Rename project: Resly -> ActiveGraph
This patch renames the project.
